Revert "Feature/19 deploy code using docker"
