+++ src/HotelManagementSystem/Admin/HotelManagementSystem.Admin.API/Program.cs
@@ -100,7 +100,7 @@ if (app.Environment.IsDevelopment())
     });
 }
 
-//app.UseHttpsRedirection();
+app.UseHttpsRedirection();
 
 app.UseCors(options => options.AllowAnyOrigin().AllowAnyHeader().AllowAnyMethod());
 app.UseAuthentication();
