+++ src/HotelManagementSystem/Admin/HotelManagementSystem.Admin.API/appsettings.json
@@ -6,9 +6,8 @@
     }
   },
   "AllowedHosts": "*",
-  "AdminConnectionString": "Server=hotelmanagementsystem.database;Database=hms_admin;User=root;Password=P@ssw0rd;Convert Zero Datetime=True;",
+  "AdminConnectionString": "Server=localhost;Database=hms_admin;User=root;Password=P@ssw0rd;Convert Zero Datetime=True;",
   "Identity": {
-    "Authorizer": "http://localhost:5001",
     "clientName": "HotelManagementSystem Administrator API",
     "clientId": "adminclient",
     "scope": "adminscope",
