+++ src/HotelManagementSystem/FrontDesk/HotelManagementSystem.FrontDesk.API/Program.cs
@@ -109,7 +109,7 @@ if (app.Environment.IsDevelopment())
     });
 }
 
-//app.UseHttpsRedirection();
+app.UseHttpsRedirection();
 
 app.UseCors(options => options.AllowAnyOrigin().AllowAnyHeader().AllowAnyMethod());
