+++ src/HotelManagementSystem/FrontDesk/HotelManagementSystem.FrontDesk.API/appsettings.json
@@ -6,10 +6,10 @@
     }
   },
   "AllowedHosts": "*",
-  "AdminConnectionString": "Server=hotelmanagementsystem.database;Database=hms_admin;User=root;Password=P@ssw0rd;Convert Zero Datetime=True;",
-  "FrontDeskConnectionString": "Server=hotelmanagementsystem.database;Database=hms_frontdesk;User=root;Password=P@ssw0rd;Convert Zero Datetime=True;",
+  "AdminConnectionString": "Server=localhost;Database=hms_admin;User=root;Password=P@ssw0rd;Convert Zero Datetime=True;",
+  "FrontDeskConnectionString": "Server=localhost;Database=hms_frontdesk;User=root;Password=P@ssw0rd;Convert Zero Datetime=True;",
   "Identity": {
-    "Authorizer": "http://localhost:5001",
+    "Authorizer": "https://localhost:7016",
     "clientName": "Front Desk Management Admin API",
     "clientId": "frontdeskapi",
     "scope": "frontdeskscope",
