+++ src/HotelManagementSystem/HotelManagementSystem.Admin.Blanket/HotelManagementSystem.Admin.Blanket.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\HotelManagementSystem.Admin.DataAccess\HotelManagementSystem.Admin.DataAccess.csproj" />
+    <ProjectReference Include="..\Admin\HotelManagementSystem.Admin.DataAccess\HotelManagementSystem.Admin.DataAccess.csproj" />
   </ItemGroup>
 
 </Project>
