+++ src/HotelManagementSystem/HotelManagementSystem.FrontDesk.Blanket/Room/RoomBlanket.cs
