+++ src/HotelManagementSystem/HotelManagementSystem.IdentityServer/Program.cs
@@ -45,12 +45,12 @@ if (!app.Environment.IsDevelopment())
     // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
     app.UseHsts();
 }
-//app.Use(async (ctx, next) =>
-//{
-//    ctx.SetIdentityServerOrigin("https://localhost:7016");
-//    await next();
-//});
-//app.UseHttpsRedirection();
+app.Use(async (ctx, next) =>
+{
+    ctx.SetIdentityServerOrigin("https://localhost:7016");
+    await next();
+});
+app.UseHttpsRedirection();
 app.UseStaticFiles();
 
 app.UseRouting();
