+++ src/HotelManagementSystem/HotelManagementSystem.IdentityServer/appsettings.json
@@ -6,7 +6,7 @@
     }
   },
   "AllowedHosts": "*",
-  "AdminConnectionString": "Server=hotelmanagementsystem.database;Database=hms_admin;User=root;Password=P@ssw0rd;Convert Zero Datetime=True;",
+  "AdminConnectionString": "Server=localhost;Database=hms_admin;User=root;Password=P@ssw0rd;Convert Zero Datetime=True;",
   "IdentityResources": [
     {
       "Enabled": true,
@@ -60,9 +60,9 @@
       "AllowedScopes": [ "adminscope", "openid", "profile" ],
       "RequirePkce": true,
       "RequireClientSecret": false,
-      "RedirectUris": [ "http://localhost:5007/login-completed", "http://localhost:5007/home"],
-      "PostLogoutRedirectUris": [ "http://localhost:5007/logout-completed" ],
-      "AllowedCorsOrigins": [ "http://localhost:5007" ],
+      "RedirectUris": [ "http://localhost:4200/login-completed", "http://localhost:4200/home"],
+      "PostLogoutRedirectUris": [ "http://localhost:4200/logout-completed" ],
+      "AllowedCorsOrigins": [ "http://localhost:4200",  ],
       "AllowOfflineAccess": true
     },
     {
