+++ src/HotelManagementSystem/HotelManagementSystem.sln
@@ -23,13 +23,11 @@ Project("{54A90642-561A-4BB1-A94E-469ADEE60C69}") = "HotelManagementSystem.Front
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HotelManagementSystem.FrontDesk.DataAccess", "FrontDesk\HotelManagementSystem.FrontDesk.DataAccess\HotelManagementSystem.FrontDesk.DataAccess.csproj", "{1541BFC8-1277-4AD2-A75B-89002F33F2F6}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HotelManagementSystem.FrontDesk.Blanket", "FrontDesk\HotelManagementSystem.FrontDesk.Blanket\HotelManagementSystem.FrontDesk.Blanket.csproj", "{4B15D745-B6B0-45F6-A674-4730174E196A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HotelManagementSystem.FrontDesk.Blanket", "HotelManagementSystem.FrontDesk.Blanket\HotelManagementSystem.FrontDesk.Blanket.csproj", "{4B15D745-B6B0-45F6-A674-4730174E196A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HotelManagementSystem.Admin.Blanket", "Admin\HotelManagementSystem.Admin.Blanket\HotelManagementSystem.Admin.Blanket.csproj", "{B39A39D5-86AA-49BF-B262-91C1F8A0B33E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HotelManagementSystem.Admin.Blanket", "HotelManagementSystem.Admin.Blanket\HotelManagementSystem.Admin.Blanket.csproj", "{B39A39D5-86AA-49BF-B262-91C1F8A0B33E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HotelManagementSystem.Test", "HotelManagementSystem.Test\HotelManagementSystem.Test.csproj", "{84D5145D-D7B7-4C56-86FB-19ACBC07D178}"
-EndProject
-Project("{E53339B2-1760-4266-BCC7-CA923CBCF16C}") = "docker-compose", "docker-compose.dcproj", "{33BD8C41-A534-400C-B0CA-DA3D6FE7CA21}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HotelManagementSystem.Test", "HotelManagementSystem.Test\HotelManagementSystem.Test.csproj", "{84D5145D-D7B7-4C56-86FB-19ACBC07D178}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -83,10 +81,6 @@ Global
 		{84D5145D-D7B7-4C56-86FB-19ACBC07D178}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{84D5145D-D7B7-4C56-86FB-19ACBC07D178}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{84D5145D-D7B7-4C56-86FB-19ACBC07D178}.Release|Any CPU.Build.0 = Release|Any CPU
-		{33BD8C41-A534-400C-B0CA-DA3D6FE7CA21}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{33BD8C41-A534-400C-B0CA-DA3D6FE7CA21}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{33BD8C41-A534-400C-B0CA-DA3D6FE7CA21}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{33BD8C41-A534-400C-B0CA-DA3D6FE7CA21}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
