--- src/HotelManagementSystem/docker-compose.override.yml
@@ -1,23 +1,0 @@
-version: '3.4'
-
-services:
-  hotelmanagementsystem.identityserver:
-    environment:
-      - ASPNETCORE_ENVIRONMENT=Development
-      - ASPNETCORE_HTTP_PORTS=5001
-    ports:
-      - "5001:5001"
-
-  hotelmanagementsystem.frontdesk.api:
-    environment:
-      - ASPNETCORE_ENVIRONMENT=Development
-      - ASPNETCORE_HTTP_PORTS=5005
-    ports:
-      - "5005:5005"
-
-  hotelmanagementsystem.admin.api:
-    environment:
-      - ASPNETCORE_ENVIRONMENT=Development
-      - ASPNETCORE_HTTP_PORTS=5003
-    ports:
-      - "5003:5003"
