--- src/HotelManagementSystem/launchSettings.json
@@ -1,11 +1,0 @@
-{
-  "profiles": {
-    "Docker Compose": {
-      "commandName": "DockerCompose",
-      "commandVersion": "1.0",
-      "serviceActions": {
-        "hotelmanagementsystem.admin.api": "StartDebugging"
-      }
-    }
-  }
-}
